fix:update get-contentdetail props connetion
DetailContent 컴포넌트로 연결되는 프롭스내용을 변경합니다.
Main.js에서 리스트를 컴포넌트화 합니다.

by @hiheecho
